Refactoring

+++ src/ffmpeg_transcoder.cc
@@ -2753,7 +2753,7 @@ void FFmpeg_Transcoder::produce_audio_dts(AVPacket *pkt, int64_t *pts)
 {
     //    if ((pkt->pts == 0 || pkt->pts == AV_NOPTS_VALUE) && pkt->dts == AV_NOPTS_VALUE)
     {
-        int duration;
+        int64_t duration;
         // Some encoders to not produce dts/pts.
         // So we make some up.
         if (pkt->duration)
